backend/auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
pytest
httpx
pytz
